requirements.txt
boto3==1.17.67
pyyaml